Add files via upload
Q-write is a method that take the key, and value from the user and adds that value to the list.. if key is already filled should return message to user to chose another key. length of list will be 10
Q-write a method with fixed length of list 5. if the user enters more the 5th value in the bottom list value should be removed and a new value should be added at the top of list
Q-write a method with list of random numbers, when user add new value method will check if the value already exist in list return message "value already in the list other wise add the value in the list and display back the list till that value
Q-write a method which can create a multi die-mention array by getting number of values of first list and number of values for the associative list, display the list as value in table formate